+++ response_writer.go
@@ -16,32 +16,6 @@ const (
 	defaultStatus = 200
 )
 
-type ResponseWriter interface {
-	http.ResponseWriter
-	http.Hijacker
-	http.Flusher
-	http.CloseNotifier
-
-	// Returns the HTTP response status code of the current request.
-	Status() int
-
-	// Returns the number of bytes already written into the response http body.
-	// See Written()
-	Size() int
-
-	// Writes the string into the response body.
-	WriteString(string) (int, error)
-
-	// Returns true if the response body was already written.
-	Written() bool
-
-	// Forces to write the http header (status code + headers).
-	WriteHeaderNow()
-
-	// get the http.Pusher for server push
-	Pusher() http.Pusher
-}
-
 type responseWriter struct {
 	http.ResponseWriter
 	size   int
@@ -114,11 +88,4 @@ func (w *responseWriter) CloseNotify() <-chan bool {
 // Flush implements the http.Flush interface.
 func (w *responseWriter) Flush() {
 	w.ResponseWriter.(http.Flusher).Flush()
-}
-
-func (w *responseWriter) Pusher() (pusher http.Pusher) {
-	if pusher, ok := w.ResponseWriter.(http.ResponseWriter).(http.Pusher); ok {
-		return pusher
-	}
-	return nil
 }
